+++ R/oeli-package.R
@@ -30,5 +30,7 @@
 #' @importFrom sysfonts font_add_google
 #' @importFrom usethis use_logo
 #' @importFrom utils combn
+#' @importFrom utils head
+#' @importFrom utils tail
 ## usethis namespace: end
 NULL